fix: map to work on insights section
